+++ src/cli/subcommands/mpool_cmd.rs
@@ -16,7 +16,7 @@ use clap::Subcommand;
 use num::BigInt;
 use std::sync::Arc;
 
-use super::{handle_rpc_err, Config};
+use super::handle_rpc_err;
 
 #[derive(Debug, Subcommand)]
 pub enum MpoolCommands {
@@ -82,12 +82,12 @@ fn filter_messages(
 async fn get_actor_sequence(
     message: &Message,
     tipset: &Arc<Tipset>,
-    config: &Config,
+    rpc_token: &Option<String>,
 ) -> Option<u64> {
     let address = message.from;
     let get_actor_result = state_get_actor(
         (address.to_owned().into(), tipset.key().to_owned().into()),
-        &config.client.rpc_token,
+        rpc_token,
     )
     .await;
 
@@ -220,7 +220,7 @@ fn print_stats(stats: &[MpStat], basefee_lookback: u32) {
 }
 
 impl MpoolCommands {
-    pub async fn run(self, config: Config) -> anyhow::Result<()> {
+    pub async fn run(self, rpc_token: Option<String>) -> anyhow::Result<()> {
         match self {
             Self::Pending {
                 local,
@@ -228,12 +228,12 @@ impl MpoolCommands {
                 to,
                 from,
             } => {
-                let messages = mpool_pending((LotusJson(vec![]),), &config.client.rpc_token)
+                let messages = mpool_pending((LotusJson(vec![]),), &rpc_token)
                     .await
                     .map_err(handle_rpc_err)?;
 
                 let local_addrs = if local {
-                    let response = wallet_list((), &config.client.rpc_token)
+                    let response = wallet_list((), &rpc_token)
                         .await
                         .map_err(handle_rpc_err)?
                         .into_inner();
@@ -262,24 +262,23 @@ impl MpoolCommands {
                 basefee_lookback,
                 local,
             } => {
-                let tipset = chain_head(&config.client.rpc_token)
+                let tipset = chain_head(&rpc_token)
                     .await
                     .map(|LotusJson(tipset)| Arc::new(tipset))
                     .map_err(handle_rpc_err)?;
                 let curr_base_fee = tipset.blocks()[0].parent_base_fee().to_owned();
 
-                let atto_str =
-                    chain_get_min_base_fee((basefee_lookback,), &config.client.rpc_token)
-                        .await
-                        .map_err(handle_rpc_err)?;
+                let atto_str = chain_get_min_base_fee((basefee_lookback,), &rpc_token)
+                    .await
+                    .map_err(handle_rpc_err)?;
                 let min_base_fee = TokenAmount::from_atto(atto_str.parse::<BigInt>()?);
 
-                let messages = mpool_pending((LotusJson(vec![]),), &config.client.rpc_token)
+                let messages = mpool_pending((LotusJson(vec![]),), &rpc_token)
                     .await
                     .map_err(handle_rpc_err)?;
 
                 let local_addrs = if local {
-                    let response = wallet_list((), &config.client.rpc_token)
+                    let response = wallet_list((), &rpc_token)
                         .await
                         .map_err(handle_rpc_err)?
                         .into_inner();
@@ -296,7 +295,7 @@ impl MpoolCommands {
 
                 let mut actor_sequences: HashMap<Address, u64> = HashMap::default();
                 for msg in messages.iter() {
-                    if let Some(sequence) = get_actor_sequence(msg, &tipset, &config).await {
+                    if let Some(sequence) = get_actor_sequence(msg, &tipset, &rpc_token).await {
                         actor_sequences.insert(msg.from, sequence);
                     }
                 }